% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colorhex_access}
\alias{colorhex_access}
\title{Check access to color-hex.com}
\usage{
colorhex_access()
}
\value{
logical if the server is accessible
}
\description{
Servers go down from time to time.
This function checks if the colorhex server
is running and retuning ok.
}
\examples{
colorhex_access()
}
