% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{get_colour}
\alias{get_colour}
\title{Get color information}
\usage{
get_colour(hex)
}
\arguments{
\item{hex}{character string that is a hexdecimal color}
}
\value{
list of class 'colorhex'
}
\description{
Get color information from www.color-hex.com
of a hex-color.
}
\examples{
get_colour("#470f0f")
get_colour("#f2f2f2")
}
