% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{get_latest_palettes}
\alias{get_latest_palettes}
\title{Get latest palettes}
\usage{
get_latest_palettes()
}
\value{
data.frame with name, id and colours
}
\description{
Retrieve the most recently made palettes
from www.color-hex.com
}
\examples{
get_latest_palettes()
}
