% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{get_palette}
\alias{get_palette}
\title{Get palettes from id}
\usage{
get_palette(id)
}
\arguments{
\item{id}{numeric id of a palette}
}
\value{
data.frame with palette information
}
\description{
Get palette information from www.color-hex.com
based on the palette id (can be found in the url)
}
\examples{
\dontrun{
get_palette(103107)

# Lookup multiple palettes
id <- c(103161, 103107)
get_palette(id)
}
}
