% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{get_popular_colors}
\alias{get_popular_colors}
\title{Get popular colour}
\usage{
get_popular_colors()
}
\value{
character vector of hex colours
}
\description{
www.color-hex.com has a list of colours
that have been liked by the most users.
This function will retrieve all of these.
}
\examples{
\dontrun{
get_popular_colors()
}
}
