% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{get_random_color}
\alias{get_random_color}
\title{Generate random HEX colour}
\usage{
get_random_color()
}
\value{
character hex value
}
\description{
Generate random HEX colour
}
\examples{
get_random_color()
}
