% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_hex}
\alias{is_hex}
\title{Validate Hex}
\usage{
is_hex(x)
}
\arguments{
\item{x}{hexidecimal character}
}
\value{
logical. TRUE if object is a hexidecimal code
}
\description{
validate if string is hexidecimal
color code
}
