% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{get_popular_palettes}
\alias{get_popular_palettes}
\title{Get most popular palettes}
\usage{
get_popular_palettes()
}
\value{
data.frame with name, id and colours
}
\description{
Retrieve the palettes most users have favorited
from www.color-hex.com
}
\examples{
get_popular_palettes()
}
